+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Introduce UnexpectedResponse exception
+
 ## v0.34.0
 
 ### Changed